Fix serenity report generation when run by multiple tags_and_report/serenity_report_when_using_tags_at_all_level
Before this change when cucumber features or scenarios are run based on logical ORs and ANDs of two or more tags, serenity report fails to create tests report. This change would fix this issue and as a result when a user runs cucumber with tags as mentioned in this link (https://github.com/cucumber/cucumber/wiki/Tags) serenity report with tests on it would be generated.

The tests are added to the maven project in the path: serenity-cucumber\src\smoketests. The feature file name is serenity_report_when_using_tags_at_all_leve.feature and Junit test name is WhenUsingScenarioOutlineAndTagsAtAllLevels.java.