Fix issue #120
Before this PR, the serenity-cucumber would generate serenity test outcome report only if the serenity-cucumber user has provided the absolute path of the feature file in the cucumber options in the JUnit tests. After this PR, the serenity-cucumber user could only provide the root folder of the feature file in cucumber options in JUnit tests and continue to reap the benefits of Serenity's test outcome report. Unit tests and integration tests are added to test this change and have been successful. Some of the existing related tests had to be refactored to support testing of this change.